Finalize the project #1 #2

+++ src/app/app.component.ts
@@ -1,22 +1,40 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { AddTodoPopupComponent } from './components/add-todo-popup/add-todo-popup.component';
+import { ConfirmDeletePopupComponent } from './components/confirm-delete-popup/confirm-delete-popup.component';
 import { Todo } from './models/todo';
 import { TodosService } from './services/todos.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
   todos$: Observable<Todo[]>;
   constructor(
     private todoService: TodosService,
-    private dialog: MatDialog,
+    private dialog: MatDialog
   ) {
-    this.todos$ = todoService.todos$;
+    this.todos$ = todoService.todos$.pipe(
+      map(todos => {
+        return todos.sort((a, b) => {
+          const aDate = new Date(a.createdAt).getMilliseconds();
+          const bDate = new Date(b.createdAt).getMilliseconds();
+          return bDate - aDate;
+        });
+      }),
+    );
+  }
+
+  ngOnInit() {
+    this.loadTodo();
+  }
+
+  loadTodo() {
+    this.todoService.getTodos();
   }
 
   addNewTodo() {
@@ -24,8 +42,8 @@ export class AppComponent {
       width: '70vw',
       disableClose: true,
     });
-    ref.afterClosed().subscribe((todo: string| null) => {
-      if(todo) {
+    ref.afterClosed().subscribe((todo: string | null) => {
+      if (todo) {
         this.todoService.addTodo(todo);
       }
     });
@@ -36,6 +54,11 @@ export class AppComponent {
   }
 
   onTodoDelete(todoId: string) {
-    this.todoService.deleteTodo(todoId);
+    this.dialog.open(ConfirmDeletePopupComponent)
+    .afterClosed().subscribe((shouldDelete) => {
+      if(shouldDelete) {
+        this.todoService.deleteTodo(todoId);
+      }
+    });
   }
 }
